@@ -2,6 +2,7 @@
   "name": "@tpzdsp/tpz-dsp3-ui-kit",
   "version": "1.0.2",
   "description": "A library template built by Vite, React, Typescript, TailwindCSS and Storybook",
+  "type": "module",
   "private": false,
   "homepage": "https://github.com/DSP3-apps/tpz-dsp3-ui-kit#readme",
   "author": "Mark Small",
@@ -43,7 +44,7 @@
     "test": "vitest",
     "test:ci": "vitest run",
     "test:e2e": "yarn playwright test",
-    "lint:rules": "eslint --print-config rules > rules.json",
+    "lint:rules": "eslint --print-config ./eslint.config.js > rules.json",
     "postinstall": "npx husky install .husky",
     "release": "semantic-release"
   },
@@ -64,26 +65,25 @@
     "@testing-library/jest-dom": "^6.4.6",
     "@testing-library/react": "^16.0.0",
     "@testing-library/user-event": "^14.5.2",
-    "@tpzdsp/dsp-tailwind-theme": "^1.1.0",
-    "@tpzdsp/eslint-config-dsp": "^1.0.1",
+    "@tpzdsp/eslint-config-dsp": "^1.2.0",
     "@types/node": "^22.13.1",
     "@types/react": "^18.2.37",
     "@types/react-dom": "^18.2.15",
-    "@typescript-eslint/eslint-plugin": "^7.16.0",
-    "@typescript-eslint/parser": "^7.16.0",
+    "@typescript-eslint/eslint-plugin": "^8.23.0",
+    "@typescript-eslint/parser": "^8.23.0",
     "@vitejs/plugin-react": "^4.3.1",
     "autoprefixer": "^10.4.16",
-    "eslint": "^8.53.0",
+    "eslint": "^9.20.0",
     "eslint-config-next": "14.2.4",
-    "eslint-config-prettier": "^9.1.0",
+    "eslint-config-prettier": "^10.0.1",
     "eslint-import-resolver-alias": "^1.1.2",
     "eslint-import-resolver-typescript": "^3.7.0",
     "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-jest-dom": "^5.4.0",
-    "eslint-plugin-jsx-a11y": "^6.9.0",
+    "eslint-plugin-jest-dom": "^5.5.0",
+    "eslint-plugin-jsx-a11y": "^6.10.2",
     "eslint-plugin-no-only-tests": "^3.1.0",
-    "eslint-plugin-prettier": "^5.1.3",
-    "eslint-plugin-react": "^7.34.3",
+    "eslint-plugin-prettier": "^5.2.3",
+    "eslint-plugin-react": "^7.37.4",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.4",
     "eslint-plugin-storybook": "^0.6.15",
@@ -99,7 +99,7 @@
     "react-dom": "^18.2.0",
     "semantic-release": "^24.0.0",
     "storybook": "^7.5.3",
-    "tailwindcss": "^3.3.5",
+    "tailwindcss": "^3.4.1",
     "typescript": "^5.5.4",
     "vite": "^5.0.0",
     "vite-plugin-dts": "^3.6.3",
@@ -107,9 +107,10 @@
   },
   "peerDependencies": {
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "tailwindcss": "^3.4.1"
   },
   "release": {
-    "extends": "./release.config.mjs"
+    "extends": "./release.config.js"
   }
 }
