@@ -1,8 +1,5 @@
-module.exports = {
-  singleQuote: true,
-  arrowParens: 'always',
-  trailingComma: 'all',
-  printWidth: 100,
-  tabWidth: 2,
-  plugins: ['prettier-plugin-tailwindcss'],
+import prettierConfig from '@tpzdsp/eslint-config-dsp/prettier/tailwind';
+
+export default {
+  ...prettierConfig,
 };
