@@ -1,3 +1,3 @@
-import './tailwind.css';
+// import './tailwind.css';
 
 export { Button } from './components/Button/Button';
