@@ -1,6 +1,5 @@
-import tailwindConfig from '@tpzdsp/dsp-tailwind-theme';
+import tailwindConfig from '@tpzdsp/eslint-config-dsp/tailwindcss';
 
-/** @type {import('tailwindcss').Config} */
 export default {
   content: ['./src/**/*.{js,ts,jsx,tsx}'],
   theme: tailwindConfig.theme,
